edit Table duzenlendi

+++ waiterApp/editTables.aspx.cs
@@ -13,19 +13,17 @@ namespace waiterApp
     {
         string connectionString = ConfigurationManager.ConnectionStrings["constring"].ConnectionString;
         fillDropDown fdp = new fillDropDown();
+        
         protected void Page_Load(object sender, EventArgs e)
         {
-            if (!Page.IsPostBack)
-            {
+            
                 Repeater1.DataSource = fdp.tableTypes(1);
                 Repeater1.DataBind();
 
-            }
         }
         protected void Repeater1_ItemDataBound(object sender, RepeaterItemEventArgs e)
         {
             SqlConnection con = new SqlConnection(connectionString);
-
             SqlCommand komut = new SqlCommand("SELECT * FROM [business].[tableinfo] ti inner join [dbo].[tableTypes] tt on ti.tableType = tt.typeID where tt.typeID=" + Convert.ToInt32(DataBinder.Eval(e.Item.DataItem, "typeID")), con);// işletmeye özel sorgu için businessID GİRİLECEK eKSİK
             con.Open();
             Repeater rp = (Repeater)e.Item.FindControl("Repeater2");
@@ -37,14 +35,34 @@ namespace waiterApp
 
         protected void Button1_Click(object sender, EventArgs e)
         {
-            SqlConnection con = new SqlConnection(connectionString);
             Button Button1 = (Button)sender;
+            int count = 0;
+            SqlConnection con = new SqlConnection(connectionString);
+            con.Open();
+            SqlCommand checkres = new SqlCommand("select count(t.tID) as resSayisi, t.tID from [business].[reservation] r inner join [business].[tableinfo] t on t.tID = r.tableID where t.tID = '"+ Button1.CommandArgument +"' group by t.tID ", con);
+            SqlDataReader dr = checkres.ExecuteReader();
+
+            while (dr.Read()) {
+               count = Convert.ToInt32(dr["resSayisi"].ToString()); 
+            }
+            con.Close();
+            
+            if(count == 0)
+            {
             con.Open();
             SqlCommand delete = new SqlCommand("delete from [business].[tableinfo] where tID = '" + Button1.CommandArgument + "'", con);
             delete.ExecuteNonQuery();
-
             con.Close();
 
+            Repeater1.DataSource = fdp.tableTypes(1);
+            Repeater1.DataBind();
+            }
+            
+
         }
-    }
+
+
+            
+        }
+    
 }
